+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Improve performance of ASN data lookups.
 * Bug: Fix a failure to save the output file when an existing file already exists.
 * Bug: Fix an issue with saving TLS details in JSON output when using `--internalssl`.
+* Bug: Fix an issue where some relative links were not being followed during the spidering process.
 
 ## 0.12.1 - 2025-04-04